+++ Assets/Scenes/SampleScene.unity
@@ -1295,6 +1295,63 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1001 &1927782054
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 3403656726215058535, guid: ff22c4f77cc992347aafc6b78dd4e041, type: 3}
+      propertyPath: m_Name
+      value: Lily
+      objectReference: {fileID: 0}
+    - target: {fileID: 7717836602779193852, guid: ff22c4f77cc992347aafc6b78dd4e041, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 2.0461826
+      objectReference: {fileID: 0}
+    - target: {fileID: 7717836602779193852, guid: ff22c4f77cc992347aafc6b78dd4e041, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0.48823088
+      objectReference: {fileID: 0}
+    - target: {fileID: 7717836602779193852, guid: ff22c4f77cc992347aafc6b78dd4e041, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 3.4737487
+      objectReference: {fileID: 0}
+    - target: {fileID: 7717836602779193852, guid: ff22c4f77cc992347aafc6b78dd4e041, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 7717836602779193852, guid: ff22c4f77cc992347aafc6b78dd4e041, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7717836602779193852, guid: ff22c4f77cc992347aafc6b78dd4e041, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7717836602779193852, guid: ff22c4f77cc992347aafc6b78dd4e041, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7717836602779193852, guid: ff22c4f77cc992347aafc6b78dd4e041, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7717836602779193852, guid: ff22c4f77cc992347aafc6b78dd4e041, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7717836602779193852, guid: ff22c4f77cc992347aafc6b78dd4e041, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: ff22c4f77cc992347aafc6b78dd4e041, type: 3}
 --- !u!1 &2087707211
 GameObject:
   m_ObjectHideFlags: 0
@@ -1591,69 +1648,11 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!1001 &3129376494606010683
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    serializedVersion: 3
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 3224206662575174608, guid: 89dd1d760450a5e429ca6ca9e3e7ef32, type: 3}
-      propertyPath: m_Name
-      value: Lily
-      objectReference: {fileID: 0}
-    - target: {fileID: 4625068583275968880, guid: 89dd1d760450a5e429ca6ca9e3e7ef32, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 2.4
-      objectReference: {fileID: 0}
-    - target: {fileID: 4625068583275968880, guid: 89dd1d760450a5e429ca6ca9e3e7ef32, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 1.14
-      objectReference: {fileID: 0}
-    - target: {fileID: 4625068583275968880, guid: 89dd1d760450a5e429ca6ca9e3e7ef32, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 3.4737487
-      objectReference: {fileID: 0}
-    - target: {fileID: 4625068583275968880, guid: 89dd1d760450a5e429ca6ca9e3e7ef32, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 4625068583275968880, guid: 89dd1d760450a5e429ca6ca9e3e7ef32, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 4625068583275968880, guid: 89dd1d760450a5e429ca6ca9e3e7ef32, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 4625068583275968880, guid: 89dd1d760450a5e429ca6ca9e3e7ef32, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 4625068583275968880, guid: 89dd1d760450a5e429ca6ca9e3e7ef32, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 4625068583275968880, guid: 89dd1d760450a5e429ca6ca9e3e7ef32, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 4625068583275968880, guid: 89dd1d760450a5e429ca6ca9e3e7ef32, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-    m_RemovedGameObjects: []
-    m_AddedGameObjects: []
-    m_AddedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 89dd1d760450a5e429ca6ca9e3e7ef32, type: 3}
 --- !u!1660057539 &9223372036854775807
 SceneRoots:
   m_ObjectHideFlags: 0
   m_Roots:
   - {fileID: 519420032}
-  - {fileID: 3129376494606010683}
   - {fileID: 1826772780}
   - {fileID: 1756992194}
   - {fileID: 2123968052}
@@ -1664,3 +1663,4 @@ SceneRoots:
   - {fileID: 2087707215}
   - {fileID: 1195748307}
   - {fileID: 1543198188}
+  - {fileID: 1927782054}
